Move functions getChapterFromReference and getVerseFromReference from AppController to VerseReferenceHelper
