+++ app/frontend/controllers/app_controller.js
@@ -692,16 +692,6 @@ class AppController {
     this.moduleAssistant.openAssistant(moduleType);
   }
 
-  getChapterFromReference(reference, separator=window.reference_separator) {
-    var chapter = Number(reference.split(separator)[0]);
-    return chapter;
-  }
-
-  getVerseFromReference(reference, separator=window.reference_separator) {
-    var verse = Number(reference.split(separator)[1]);
-    return verse;
-  }
-
   getBibleBookStatsFromVerseList(tabIndex) {
     var bibleBookStats = {};    
     var currentVerseList = this.getCurrentVerseList(tabIndex)[0];
